fix:comment

+++ src/main/java/hackathon/bigone/sunsak/groupbuy/comment/controller/GroupBuyCommentController.java
@@ -1,6 +1,7 @@
 package hackathon.bigone.sunsak.groupbuy.comment.controller;
 
 import hackathon.bigone.sunsak.accounts.user.entity.SiteUser;
+import hackathon.bigone.sunsak.global.security.jwt.CustomUserDetail;
 import hackathon.bigone.sunsak.groupbuy.comment.dto.GroupBuyCommentRequestDto;
 import hackathon.bigone.sunsak.groupbuy.comment.dto.GroupBuyCommentResponseDto;
 import hackathon.bigone.sunsak.groupbuy.comment.service.GroupBuyCommentService;
@@ -26,19 +27,23 @@ public class GroupBuyCommentController {
     public ResponseEntity<GroupBuyCommentResponseDto> addComment(
             @PathVariable Long groupbuyId,
             @RequestBody @Valid GroupBuyCommentRequestDto requestDto,
-            @AuthenticationPrincipal SiteUser user) {
+            @AuthenticationPrincipal CustomUserDetail userDetail) {
 
+        if (userDetail == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
+        SiteUser user = userDetail.getUser();
         GroupBuyCommentResponseDto newComment = GroupBuyCommentService.addComment(groupbuyId, requestDto, user);
-        return ResponseEntity.status(HttpStatus.CREATED).body(newComment);
+        return new ResponseEntity<>(newComment, HttpStatus.CREATED);
     }
 
     //댓글 조회
     @GetMapping
-    public ResponseEntity<List<GroupBuyCommentResponseDto>> getComments(
-            @PathVariable Long groupbuyId) {
+    public List<GroupBuyCommentResponseDto> getComments(
+                                                         @PathVariable Long groupbuyId) {
 
-        List<GroupBuyCommentResponseDto> comments = GroupBuyCommentService.getComments(groupbuyId);
-        return ResponseEntity.ok(comments);
+        return GroupBuyCommentService.getComments(groupbuyId);
     }
 
     //댓글 삭제
@@ -46,9 +51,14 @@ public class GroupBuyCommentController {
     public ResponseEntity<Void> deleteComment(
             @PathVariable Long groupbuyId,
             @PathVariable Long commentId,
-            @AuthenticationPrincipal SiteUser user) {
+            @AuthenticationPrincipal CustomUserDetail userDetail) {
 
+        if (userDetail == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
+        SiteUser user = userDetail.getUser();
         GroupBuyCommentService.deleteComment(groupbuyId, commentId, user);
-        return ResponseEntity.noContent().build();
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
